+++ README.Rmd
@@ -17,6 +17,7 @@ knitr::opts_chunk$set(
 
 <!-- badges: start -->
 [![Codecov test coverage](https://codecov.io/gh/pabl1t0x/SatRbedo/branch/master/graph/badge.svg)](https://app.codecov.io/gh/pabl1t0x/SatRbedo/tree/master)
+[![status](https://joss.theoj.org/papers/c5899414306d4cf318350654b7aa38fa/status.svg)](https://joss.theoj.org/papers/c5899414306d4cf318350654b7aa38fa)
 <!-- badges: end -->
 
 The `SatRbedo` package consists of a set of tools for retrieving snow and ice albedo from optical satellite imagery (e.g., [Landsat](https://www.usgs.gov/landsat-missions/landsat-collection-2-surface-reflectance) and [Sentinel-2](https://dataspace.copernicus.eu/explore-data/data-collections/sentinel-data/sentinel-2)). `SatRbedo` functions require the following datasets:
