### 2025-07-22 #8429 @kthyng
@pabl1t0x
* Can you host the docs somewhere so a user that is considering your package online can browse them without having to install the package?
