+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-url: ~
+url: https://pabl1t0x.github.io/SatRbedo/
 template:
   bootstrap: 5
 
