+++ pkgdown.yml
@@ -1,0 +1,8 @@
+pandoc: 3.1.11
+pkgdown: 2.1.3
+pkgdown_sha: ~
+articles: {}
+last_built: 2025-08-01T20:29Z
+urls:
+  reference: https://pabl1t0x.github.io/SatRbedo/reference
+  article: https://pabl1t0x.github.io/SatRbedo/articles
