### 2025-08-01 #8429 @pabl1t0x
Hi @kthyng.

I have deployed a pkgdown website that hosts the docs. I added the website to the SatRbedo Github repository (https://pabl1t0x.github.io/SatRbedo/).
