### 2025-09-02 #8429 @ethanwhite
Thanks @trchudley @pritamd47 and @PennyHow and apologies for forgetting to check for language compatibility as well as subject compatibility.

@pabl1t0x - I'll be inviting another round of reviews shortly.
