### 2025-09-08 #8429 @ethanwhite
👋 @DS4Ag @dlebauer @stevenpawley  - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
