### 2025-09-11 #8973 @editorialbot
[REVIEW]: SatRbedo: An R package for retrieving snow and ice albedo from optical satellite imagery
**Submitting author:** @pabl1t0x (<a href="http://orcid.org/0000-0002-6042-5620">Pablo Fuchs</a>)
**Repository:** https://github.com/pabl1t0x/SatRbedo
**Branch with paper.md** (empty if default branch):
**Version:** v1.0

## Editor Notes

- @dlebauer won't be submitting review until mid-late October
