### 2025-09-26 #8973 @ethanwhite
👋 @DS4Ag. Thanks again for agreeing to review this paper. I can see that you've already started your checklist. Much appreciated! Since we're now two weeks into the review process, I just wanted to check in and see if you have any questions or have run into any issues that I can help out with as you work on your reviews.

 @dlebauer - I know I'm not expecting anything from you for a few weeks.
