### 2025-10-01 #8973 @DS4Ag
I have started the review; if it's okay with @ethanwhite, I will place the code-related comments in an issue in the [SatRbedo repository](https://github.com/pabl1t0x/SatRbedo).

However, I think this comment related to the license is appropriate here:

The repository includes the full GPLv3 text, which is correct, but it has not been adapted to this project. The placeholders in the “How to Apply These Terms…” section (e.g., &lt;program name&gt;, &lt;year&gt;, &lt;author&gt;) need to be replaced with project-specific information.
