### 2025-10-02 #8973 @ethanwhite
Thanks @DS4Ag! And yes, generally comments are best in the SatRbedo repo and also the note here on the license is fine.
