### 2025-10-15 pabl1t0x/SatRbedo#1 @dlebauer
JOSS Manuscript feedback
These are notes from my first read-through (for https://github.com/openjournals/joss-reviews/issues/8973)

- [x] line 59: Since two specific cases are cited, I think it would be helpful to be be more specific than "not always available" in order to point out the limitations of the status quo. With respect to the two cited papers, what is the actual status of the code? available? on request? Did you contact the authors of the cited software? This could be an additional clause or sentence.
- [x] Figure 1: is complex but helpful. I think the following could make it more clear to the reader
  - [x] on the top line, it is difficult to tell the arrow direction because the arrow is too small and indistinguishable from the starting ball.
  - [x] could you add function names to the diagram or legend to help link the workflow illustration to the software components?
  - [x] I don't think the enumerations in the legend is necessary. First, it suggests there should be numbers on the figure. Second, there are already man distinct enumerated lists in the prose (lines 50-51, 66-68, 82-84).
  - [x] But ... it would help to clearly identify each processing step by name in the figure. Use consistent names and capitalize each step. For example:
    - [x] on line 63 count the distinct tools (four) and use "workflow" instead of "tools" for consistency. "SatRbedo consists of **four** ~tools~ **workflows** ...",
    - [x] line 64 where it says "First, it takes..." You might say "First, ~it~ **the Preprocessing Workflow** ... " and on line 70 "Subsequently, ~it~ **the Topographic Correction workflow**
  - [x] Please make sure all symbols in the figure are defined (I don't see where $\alpha$ is defined).
- [x] line 58 ~worflows~ --> workflows
- [x] line 78 typo ~cxast~ --> cast
- [x] line 99: cite the terra package `citation("terra")`
- [x] line 100: ~Corripi~ --> Corripio
