### 2025-10-15 pabl1t0x/SatRbedo#2 @dlebauer
Installation of Bioconductor dependencies
JOSS review https://github.com/openjournals/joss-reviews/issues/8973

SatRbedo uses the bioconductor package EBImage

When I tried to install following the README, I get the error:

> dependency ‘EBImage’ is not available for package ‘SatRbedo’

Adding the following to the DESCRIPTION should automate installation of EBImage when using `remotes::install_github()` and `pad::pak()`, (but not `install.packages()`):
```
Additional_repositories:
    https://bioconductor.org/packages/release/bioc
Remotes:
    bioc::EBImage
```

Alternatively, add installation instructions to the README:

```
if (!requireNamespace("BiocManager", quietly = TRUE))
    install.packages("BiocManager")
BiocManager::install("pabl1t0x/SatRbedo")
```
