### 2025-10-16 pabl1t0x/SatRbedo#3 @dlebauer
Add community guidelines
re: https://github.com/openjournals/joss-reviews/issues/8973

From JOSS review criteria see [Community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) should include

"clear guidelines for third-parties wishing to: Contribute to the software, Report issues or problems with the software, or Seek support"

This can be done in the README or a new CONTRIBUTING.md file. If it is in a CONTRIBUTING.md file, I think a link will show up when contributors create a new issue or PR.
