### 2025-10-16 pabl1t0x/SatRbedo#4 @dlebauer
Remove pre-compiled binaries from repository
re: https://github.com/openjournals/joss-reviews/issues/8973

Precompiled binary files (`*.o`, `*.so`) are platform specific and generated during installation. When I tried to install this package on MacOS I got errors including "no such symbol doshade".

**Suggestion:**

1. remove these files:

- src/SatRbedo.so
- src/cast_shadows.o
- src/init.o

2. add the following to .gitignore:

```
src/*.o
src/*.so
```
