+++ R/shadow_removal.R
@@ -1,3 +1,19 @@
+# SatRbedo: Tools for retrieving snow and ice albedo from optical satellite imagery.
+# Copyright (C) 2025  Pablo Fuchs
+
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
 #' Topographic shadows
 #'
 #' Detects and removes self-shadowed pixels (i.e., pixels having
