+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.3
 pkgdown_sha: ~
 articles: {}
-last_built: 2025-08-01T20:29Z
+last_built: 2025-10-18T20:39Z
 urls:
   reference: https://pabl1t0x.github.io/SatRbedo/reference
   article: https://pabl1t0x.github.io/SatRbedo/articles
