### 2025-10-19 dfc4dba
Deploying to gh-pages from @ pabl1t0x/SatRbedo@b202de52ee58e760b576f38f6451da7e1e09a9ce 🚀
