### 2025-10-21 a51b901
Create CODE_OF_CONDUCT.md
