### 2025-10-21 0b821ed
Create CONTRIBUTING.md
