+++ CONTRIBUTING.md
@@ -1,0 +1,21 @@
+# Contributing
+Thank you for your interest in contributing to SatRbedo! To ensure a smooth experience, please review the guidelines provided here.
+## Community Code of Conduct
+By participating in the SatRbedo community, you are expected to uphold this [Code of Conduct](CODE_OF_CONDUCT.md), which defines the standards
+for a positive and constructive environment.
+## How to contribute
+You can contribute to SatRbedo by:
+- Reporting or fixing an [issue](https://github.com/pabl1t0x/SatRbedo/issues).
+- Reviewing a [Pull Request](https://github.com/pabl1t0x/SatRbedo/pulls).
+- Suggesting new features.
+- Updating the [package](README.Rmd) and [function](/man) documentation.
+- Writing or updating [test functions](/tests).
+- Writing tutorials (vignettes).
+
+For all contributions, please respect the following guidelines:
+- Check your spelling and grammar.
+- Search previous suggestions before making a new one, as yours may be a duplicate.
+- Make an individual pull request for each suggestion.
+- The pull request and commit should have a useful title.
+- Be open to criticism to improve your code.
+- Please use the [Tydiverse](https://style.tidyverse.org) and [Google's](https://google.github.io/styleguide/Rguide.html) R style guide.
