+++ CONTRIBUTING.md
@@ -1,9 +1,15 @@
 # Contributing
+
 Thank you for your interest in contributing to `SatRbedo`! To ensure a smooth experience, please review the guidelines provided here.
+
 ## Community Code of Conduct
+
 By participating in the `SatRbedo` community, you are expected to uphold this [Code of Conduct](CODE_OF_CONDUCT.md), which defines the standards for a positive and constructive environment.
+
 ## How to contribute
+
 You can contribute to `SatRbedo` by:
+
 - Reporting or fixing an [issue](https://github.com/pabl1t0x/SatRbedo/issues).
 - Reviewing a [Pull Request](https://github.com/pabl1t0x/SatRbedo/pulls).
 - Suggesting new features.
@@ -12,6 +18,7 @@ You can contribute to `SatRbedo` by:
 - Writing tutorials (vignettes).
 
 For all contributions, please respect the following guidelines:
+
 - Check your spelling and grammar.
 - Search previous suggestions before making a new one, as yours may be a duplicate.
 - Make an individual pull request for each suggestion.
