+++ DESCRIPTION
@@ -73,7 +73,7 @@ Additional_repositories:
 Remotes:
     bioc::EBImage
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
 Suggests: 
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
