+++ DESCRIPTION
@@ -75,7 +75,8 @@ Remotes:
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.3
 Suggests: 
-    testthat (>= 3.0.0)
+    testthat (>= 3.0.0),
+    vdiffr
 Config/testthat/edition: 3
 URL: https://github.com/pabl1t0x/SatRbedo, https://pabl1t0x.github.io/SatRbedo/
 BugReports: https://github.com/pabl1t0x/SatRbedo/issues
