### 2025-10-21 #8973 @pabl1t0x
> I have started the review; if it's okay with [@ethanwhite](https://github.com/ethanwhite), I will place the code-related comments in an issue in the [SatRbedo repository](https://github.com/pabl1t0x/SatRbedo).
>
> However, I think these comments are appropriate here:
>
> 1. The repository includes the full GPLv3 text, which is correct, but it has not been adapted to this project. The placeholders in the “How to Apply These Terms…” section (e.g., <program name>, <year>, <author>) need to be replaced with project-specific information.
> 2. The repository does not have "Community guidelines".
> 3. Expand the test coverage on the following functions:
>
> * shadow_removal.R
> * snow_or_ice.R
> * shadow_removal.R

Hi @DS4Ag,

Thanks for your feedback. I have now addressed your comments:

1. The placeholders in the "How to Apply These Terms" [section](https://github.com/pabl1t0x/SatRbedo/blob/master/LICENSE.md) have been updated, as suggested. I have also added a short license notice at the start of each source file in the [R](https://github.com/pabl1t0x/SatRbedo/tree/master/R) directory.
2. The repository now includes [Community Guidelines](https://github.com/pabl1t0x/SatRbedo/blob/master/CONTRIBUTING.md).
3. I created new tests for the shadow_removal.R and snow_or_ice.R functions. You can find them [here](https://github.com/pabl1t0x/SatRbedo/tree/master/tests/testthat).

Please let me know if you have further comments.
