### 2025-10-22 884ca5f
Revised JOSS manuscript
