+++ inst/paper/paper.md
@@ -70,7 +70,7 @@ Coupled snow radiative transfer and snowpack models can simulate the temporal an
 Remote sensing offers the best option for studying the changes in albedo, accounting for the high variability in space and time [@berg_finescale_2020]. Satellite albedo retrievals typically comprise three steps: (1) atmospheric correction, (2) modeling of the angular reflectance, and (3) narrow-to-broadband albedo conversion [@carlsen_parameterizing_2020;@qu_mapping_2015]. The algorithms for atmospheric correction [@doxani_atmospheric_2023;@vermote_preliminary_2016], modeling of the angular reflectance distribution [@lucht_algorithm_2000;@ren_anisotropy_2021], and narrow-to-broadband albedo conversion
 [@knap_narrowband_1999;@li_preliminary_2018;@liang_narrowband_2001] are well-established and validated in a number of case studies. In addition to these steps, satellite image pre-processing and topographic correction are necessary for homogenizing the input data and minimizing the effects of slope and aspect on albedo, respectively.
 
-Several worflows have been proposed to address these processing steps [e.g.,@klok_temporal_2003;@shuai_algorithm_2011]. However, research code supporting their implementation is not always readily available. This paper introduces the `SatRbedo` package, which implements a workflow in R to estimate snow and ice albedo from satellite data.
+Several workflows have been proposed to address these processing steps [e.g.,@klok_temporal_2003;@shuai_algorithm_2011]. However, research code supporting their implementation is not always readily available. This paper introduces the `SatRbedo` package, which implements a workflow in R to estimate snow and ice albedo from satellite data.
 
 # Implementation
 
@@ -80,7 +80,7 @@ Several worflows have been proposed to address these processing steps [e.g.,@klo
 
 First, it takes application-ready surface reflectance data ($\rho_s$) derived from top-of-atmosphere reflectance, satellite ($\varphi_v$, $\theta_v$) and solar ($\varphi_s$, $\theta_s$) azimuth and zenith angles, and an outline of the area of interest (AOI) to perform the following pre-processing steps: (1) crop the satellite grids to a specified extent; (2) convert data from integer to floating point; (3) re-project grids to a common coordinate system; and (4) convert nadir satellite observations to off-nadir values using view-angle corrections based on the c-factor method [@roy_general_2016].
 
-Subsequently, the non-view-angle corrected surface reflectance ($\rho_T$) is corrected for the effects of topography to obtain the equivalent reflectance values over flat terrain ($\rho_H$). Two empirical methods are provided in `SatRbedo`: (1) the rotation model proposed by @tan_improved_2013 and the C-correction model [@teillet_slope-aspect_1982]. These algorithms are suitable for mountain environments with rugged topography and non-Lambertian surface properties and require a digital elevation model (DEM) and the solar azimuth and zenith angles as input data. The two models assume a linear relationship between reflectance and the solar incidence angle on an inclined surface. Additionally, a tool is provided to remove topographic shadows (self and cxast shadows) using the vectorial algebra algorithms proposed by @corripio_vectorial_2003.
[…]
