+++ inst/paper/paper.bib
@@ -152,6 +152,18 @@
 	pages = {265--277},
 }
 
+@misc{hijmans_terra_2025,
+	title = {terra: {Spatial} {Data} {Analysis}},
+	shorttitle = {terra},
+	url = {https://rspatial.org},
+	doi = {10.32614/CRAN.package.terra},
+	language = {en},
+	urldate = {2025-10-22},
+	author = {Hijmans, Robert J.},
+	year = {2025},
+	note = {R package version 1.8-74},
+}
+
 @article{hock_glacier_2005,
 	title = {Glacier melt: a review of processes and their modelling},
 	volume = {29},
