### 2025-10-23 95ec5b9
Revised JOSS manuscript
