+++ inst/paper/paper.bib
@@ -152,16 +152,12 @@
 	pages = {265--277},
 }
 
-@misc{hijmans_terra_2025,
-	title = {terra: {Spatial} {Data} {Analysis}},
-	shorttitle = {terra},
-	url = {https://rspatial.org},
-	doi = {10.32614/CRAN.package.terra},
-	language = {en},
-	urldate = {2025-10-22},
-	author = {Hijmans, Robert J.},
-	year = {2025},
-	note = {R package version 1.8-74},
+@Manual{,
+  title = {terra: Spatial Data Analysis},
+  author = {Robert J. Hijmans},
+  year = {2025},
+  note = {R package version 1.8-74},
+  url = {https://rspatial.org/},
 }
 
 @article{hock_glacier_2005,
