### 2025-10-23 0c41c7d
Revised manuscript
