+++ inst/paper/paper.md
@@ -67,7 +67,7 @@ Instrumentation deployed on towers and automatic weather stations can provide hi
 
 Coupled snow radiative transfer and snowpack models can simulate the temporal and spatial evolution of snow optical properties. However, these coupled models are associated with high computational costs and limited spatial resolution [@gaillard_improving_2025]. Also, these models require input data that is often spatially variable and unavailable most of the time.
 
-Remote sensing offers the best option for studying the changes in albedo, accounting for the high variability in space and time [@berg_finescale_2020]. Satellite albedo retrievals typically comprise three steps: (1) atmospheric correction, (2) modeling of the angular reflectance, and (3) narrow-to-broadband albedo conversion [@carlsen_parameterizing_2020;@qu_mapping_2015]. The algorithms for atmospheric correction [@doxani_atmospheric_2023;@vermote_preliminary_2016], modeling of the angular reflectance distribution [@lucht_algorithm_2000;@ren_anisotropy_2021], and narrow-to-broadband albedo conversion
+Remote sensing offers the best option for studying the changes in albedo, accounting for the high variability in space and time [@berg_finescale_2020]. Satellite albedo retrievals typically comprise three steps: atmospheric correction, modeling of the angular reflectance, and narrow-to-broadband albedo conversion [@carlsen_parameterizing_2020;@qu_mapping_2015]. The algorithms for atmospheric correction [@doxani_atmospheric_2023;@vermote_preliminary_2016], modeling of the angular reflectance distribution [@lucht_algorithm_2000;@ren_anisotropy_2021], and narrow-to-broadband albedo conversion
 [@knap_narrowband_1999;@li_preliminary_2018;@liang_narrowband_2001] are well-established and validated in a number of case studies. In addition to these steps, satellite image pre-processing and topographic correction are necessary for homogenizing the input data and minimizing the effects of slope and aspect on albedo, respectively.
 
 Several workflows have been proposed to address these processing steps [e.g.,@klok_temporal_2003;@shuai_algorithm_2011]. However, research code supporting their implementation is not always readily available. This paper introduces the `SatRbedo` package, which implements a workflow in R to estimate snow and ice albedo from satellite data.
@@ -76,13 +76,13 @@ Several workflows have been proposed to address these processing steps [e.g.,@kl
 
 `SatRbedo` consists of tools that run in a processing pipeline (\autoref{fig:workflow}).
 
-![Flowchart of the satellite albedo retrieval workflow. It includes four processing steps: (1) pre-processing, (2) topographic correction, (3) anisotropic correction, and (4) narrow-to-broadband albedo conversion. The details of the methods are described in the text.\label{fig:workflow}](workflow.png){ width=80% }
+![Flowchart of the satellite albedo retrieval workflow. It includes four processing steps: pre-processing, topographic correction, anisotropic correction, and narrow-to-broadband albedo conversion. The details of the methods are described in the text.\label{fig:workflow}](workflow.png){ width=80% }
 
-First, it takes application-ready surface reflectance data ($\rho_s$) derived from top-of-atmosphere reflectance, satellite ($\varphi_v$, $\theta_v$) and solar ($\varphi_s$, $\theta_s$) azimuth and zenith angles, and an outline of the area of interest (AOI) to perform the following pre-processing steps: (1) crop the satellite grids to a specified extent; (2) convert data from integer to floating point; (3) re-project grids to a common coordinate system; and (4) convert nadir satellite observations to off-nadir values using view-angle corrections based on the c-factor method [@roy_general_2016].
[…]
