+++ inst/paper/paper.bib
@@ -152,7 +152,7 @@
 	pages = {265--277},
 }
 
-@Manual{,
+@Manual{hijmans_terra_2025,
   title = {terra: Spatial Data Analysis},
   author = {Robert J. Hijmans},
   year = {2025},
