### 2025-10-25 c219386
Revised manuscript
