+++ inst/paper/paper.md
@@ -43,12 +43,9 @@ snow and ice surfaces. As such, it influences meltwater production, glacier mass
 balance, and the energy exchange between the Earth and the atmosphere [@jonsell_spatial_2003;@hock_glacier_2005]. Satellite remote sensing has been widely recognized as the best practical
 approach for monitoring and mapping surface albedo across different spatial and temporal scales [@lin_estimating_2022;@urraca_assessing_2023]. Here, we present the `SatRbedo` R package:
 an extensible, standalone toolbox for retrieving snow and ice albedo from optical satellite imagery.
-The package includes tools for image preprocessing, converting nadir satellite observations to off-nadir values using view-angle corrections, detecting topographic shadows, discriminating snow and ice surfaces, correcting for topographic effects and the anisotropic behavior of reflected radiation of glacier snow
-and ice, and converting narrowband to broadband albedo. The toolbox has a modular structure that
-allows for changing the implemented routines and provides output that can be used independently
-or as input to other functions. `SatRbedo` is designed to work with medium-resolution
-satellite data (e.g., Landsat and Sentinel-2), although data from different satellite sensors
-can also be used.
+The package includes functions for image preprocessing, converting nadir satellite observations to off-nadir values using view-angle corrections, detecting topographic shadows, discriminating snow and ice surfaces, correcting for topographic effects and the anisotropic behavior of reflected radiation of glacier snow and ice, and converting narrowband to broadband albedo. The package has a modular structure that allows for changing the implemented routines and provides output that can be used independently
+or as input to other functions. `SatRbedo` is designed to work with medium-resolution satellite data
+(e.g., Landsat and Sentinel-2), although data from different satellite sensors can also be used.
 
 # Statement of need
 
@@ -68,25 +65,24 @@ Instrumentation deployed on towers and automatic weather stations can provide hi
 Coupled snow radiative transfer and snowpack models can simulate the temporal and spatial evolution of snow optical properties. However, these coupled models are associated with high computational costs and limited spatial resolution [@gaillard_improving_2025]. Also, these models require input data that is often spatially variable and unavailable most of the time.
 
 Remote sensing offers the best option for studying the changes in albedo, accounting for the high variability in space and time [@berg_finescale_2020]. Satellite albedo retrievals typically comprise three steps: atmospheric correction, modeling of the angular reflectance, and narrow-to-broadband albedo conversion [@carlsen_parameterizing_2020;@qu_mapping_2015]. The algorithms for atmospheric correction [@doxani_atmospheric_2023;@vermote_preliminary_2016], modeling of the angular reflectance distribution [@lucht_algorithm_2000;@ren_anisotropy_2021], and narrow-to-broadband albedo conversion
-[@knap_narrowband_1999;@li_preliminary_2018;@liang_narrowband_2001] are well-established and validated in a number of case studies. In addition to these steps, satellite image pre-processing and topographic correction are necessary for homogenizing the input data and minimizing the effects of slope and aspect on albedo, respectively.
+[@knap_narrowband_1999;@li_preliminary_2018;@liang_narrowband_2001] are well-established and validated in a number of case studies. In addition to these steps, satellite image preprocessing and topographic correction are necessary for homogenizing the input data and minimizing the effects of slope and aspect on albedo, respectively.
 
-Several workflows have been proposed to address these processing steps [e.g.,@klok_temporal_2003;@shuai_algorithm_2011]. However, research code supporting their implementation is not always readily available. This paper introduces the `SatRbedo` package, which implements a workflow in R to estimate snow and ice albedo from satellite data.
[…]
