+++ README.Rmd
@@ -9,10 +9,10 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   fig.path = "./man/figures/",
-  out.width = "90%"
+  out.width = "100%"
 )
-knitr::opts_knit$set(global.par = TRUE)
 ```
+<!-- knitr::opts_knit$set(global.par = TRUE) -->
 
 # SatRbedo
 
@@ -52,16 +52,14 @@ devtools::install_github("pabl1t0x/SatRbedo")
 Snow and ice albedo retrieval from satellite imagery is a four-step workflow that includes image pre-processing, topographic and anisotropic corrections, and a narrow-to-broadband albedo conversion (Fig. 1).
 
 ``` {r workflow, echo = FALSE, fig.cap = "Fig. 1 Flowchart of the satellite albedo retrieval workflow"}
-knitr::include_graphics("./man/figures/workflow.png")
+knitr::include_graphics("man/figures/workflow.png")
 ```
 
 The following example shows how to estimate albedo using surface reflectance data from five spectral bands at the Athabasca Glacier in Canada on 11 September 2020.
 
 **Step 1: Load the data for the area of interest**
 
-```{r}
-par(mar=c(1,1,1,1))
-``` 
+<!-- par(mar=c(1,1,1,1)) -->
 
 ```{r example-load}
 # Load the packages
