+++ README.Rmd
@@ -9,7 +9,7 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   fig.path = "./man/figures/",
-  out.width = "100%"
+  out.width = "80%"
 )
 ```
 <!-- knitr::opts_knit$set(global.par = TRUE) -->
