+++ README.Rmd
@@ -8,7 +8,7 @@ output: github_document
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  fig.path = "./man/figures/",
+  fig.path = "man/figures/",
   out.width = "80%"
 )
 ```
