### 2025-10-27 ff28127
Revised paper
