+++ inst/paper/paper.md
@@ -64,10 +64,10 @@ Instrumentation deployed on towers and automatic weather stations can provide hi
 
 Coupled snow radiative transfer and snowpack models can simulate the temporal and spatial evolution of snow optical properties. However, these coupled models are associated with high computational costs and limited spatial resolution [@gaillard_improving_2025]. Also, these models require input data that is often spatially variable and unavailable most of the time.
 
-Remote sensing offers the best option for studying the changes in albedo, accounting for the high variability in space and time [@berg_finescale_2020]. Satellite albedo retrievals typically comprise three steps: atmospheric correction, modeling of the angular reflectance, and narrow-to-broadband albedo conversion [@carlsen_parameterizing_2020;@qu_mapping_2015]. The algorithms for atmospheric correction [@doxani_atmospheric_2023;@vermote_preliminary_2016], modeling of the angular reflectance distribution [@lucht_algorithm_2000;@ren_anisotropy_2021], and narrow-to-broadband albedo conversion
+Remote sensing offers the best option for studying the changes in albedo, accounting for the high variability in space and time [@berg_finescale_2020]. Satellite albedo retrievals typically comprise three steps: atmospheric correction, modeling of the angular reflectance, and narrow-to-broadband albedo conversion [@carlsen_parameterizing_2020;@qu_mapping_2015]. The algorithms for atmospheric correction [@doxani_atmospheric_2023;@vermote_preliminary_2016], modeling of the angular reflectance distribution [@klok_temporal_2003;@lucht_algorithm_2000;@ren_anisotropy_2021], and narrow-to-broadband albedo conversion
 [@knap_narrowband_1999;@li_preliminary_2018;@liang_narrowband_2001] are well-established and validated in a number of case studies. In addition to these steps, satellite image preprocessing and topographic correction are necessary for homogenizing the input data and minimizing the effects of slope and aspect on albedo, respectively.
 
-Several workflows have been proposed to address these processing steps [e.g.,@klok_temporal_2003;@shuai_algorithm_2011]. However, research code supporting their implementation is not always readily available. This paper introduces the `SatRbedo` package, which implements workflows in R to estimate snow and ice albedo from satellite data.
+Although many useful workflows are available to retrieve snow and ice albedo [@bertoncini_large-area_2022;@feng_remote_2023;@mullen_operational_2022;@ren_observed_2024], these workflows are often designed for specific case studies, have limited modularity and documentation, and lack function testing. This restricts their applicability in other regions and their long-term maintainability. This paper introduces `SatRbedo`, an extensible and well-documented R package that provides tools for estimating snow and ice albedo from satellite data.
 
 # Implementation
 
