+++ inst/paper/paper.bib
@@ -266,6 +266,16 @@
 	pages = {977--998},
 }
 
+@article{mullen_operational_2022,
+	title = {An {Operational} {Methodology} for {Validating} {Satellite}-{Based} {Snow} {Albedo} {Measurements} {Using} a {UAV}},
+	volume = {2},
+	doi = {10.3389/frsen.2021.767593},
+	journal = {Frontiers in Remote Sensing},
+	author = {Mullen, Andrew and Sproles, Eric A. and Hendrikx, Jordy and Shaw, Joseph A. and Gatebe, Charles K.},
+	year = {2022},
+	pages = {767593},
+}
+
 @article{otsu_threshold_1979,
 	title = {A {Threshold} {Selection} {Method} from {Gray}-{Level} {Histograms}},
 	volume = {9},
@@ -321,6 +331,18 @@
 	pages = {1714},
 }
 
+@article{ren_observed_2024,
+	title = {Observed and projected declines in glacier albedo across the {Third} {Pole} in the 21st century},
+	volume = {7},
+	doi = {10.1016/j.oneear.2024.08.010},
+	number = {9},
+	urldate = {2025-10-27},
+	journal = {One Earth},
+	author = {Ren, Shaoting and Jia, Li and Miles, Evan S. and Menenti, Massimo and Kneib, Marin and Shaw, Thomas E. and Buri, Pascal and McCarthy, Michael J. and Yang, Wei and Pellicciotti, Francesca and Yao, Tandong},
+	year = {2024},
+	pages = {1587--1599},
+}
+
 @article{roy_general_2016,
 	title = {A general method to normalize {Landsat} reflectance data to nadir {BRDF} adjusted reflectance},
 	volume = {176},
@@ -340,17 +362,6 @@
 	author = {Schaepman-Strub, G. and Schaepman, M.E. and Painter, T.H. and Dangel, S. and Martonchik, J.V.},
 	year = {2006},
 	pages = {27--42},
-}
-
-@article{shuai_algorithm_2011,
-	title = {An algorithm for the retrieval of 30-m snow-free albedo from {Landsat} surface reflectance and {MODIS} {BRDF}},
-	volume = {115},
-	doi = {10.1016/j.rse.2011.04.019},
-	number = {9},
-	journal = {Remote Sensing of Environment},
-	author = {Shuai, Yanmin and Masek, Jeffrey G. and Gao, Feng and Schaaf, Crystal B.},
-	year = {2011},
-	pages = {2204--2216},
 }
 
 @article{tan_improved_2013,
