### 2025-10-28 pabl1t0x/SatRbedo#1 @pabl1t0x
Thanks for your feedback on the manuscript, @dlebauer. You can find my responses to your comments below:

**line 59: Since two specific cases are cited, I think it would be helpful to be more specific than "not always available" in order to point out the limitations of the status quo. With respect to the two cited papers, what is the actual status of the code? available? on request? Did you contact the authors of the cited software? This could be an additional clause or sentence.**

Thanks for the suggestion. I wrote a new paragraph that includes specific information about open-access tools and their limitations.

**Figure 1: is complex but helpful. I think the following could make it more clear to the reader**
**- on the top line, it is difficult to tell the arrow direction because the arrow is too small and indistinguishable from the starting ball.**

Figure 1 has been modified, based on your feedback. The arrow size has been increased for clarity.

**- could you add function names to the diagram or legend to help link the workflow illustration to the software components?**

Function names have been added to the diagram and legend.

**- I don't think the enumerations in the legend is necessary. First, it suggests there should be numbers on the figure. Second, there are already man distinct enumerated lists in the prose (lines 50-51, 66-68, 82-84).**

I have removed enumerations from both the text and the figure caption.

**- But ... it would help to clearly identify each processing step by name in the figure. Use consistent names and capitalize each step. For example:**
    **- on line 63 count the distinct tools (four) and use "workflow" instead of "tools" for consistency. "SatRbedo consists of four ~tools~ workflows ...",**
    **- line 64 where it says "First, it takes..." You might say "First, ~it~ the Preprocessing Workflow ... " and on line 70 "Subsequently, ~it~ the Topographic Correction workflow**

To improve clarity, I have edited the text as suggested.

**Please make sure all symbols in the figure are defined (I don't see where $\alpha$ is defined).**

I have verified that all symbols in the figure are defined.

**line 58 ~worflows~ --> workflows**

Changed accordingly.

**line 78 typo ~cxast~ --> cast**

Changed accordingly.

**line 99: cite the terra package `citation("terra")`**

Included a citation for the terra package.

**line 100: ~Corripi~ --> Corripio**

Changed accordingly.
