### 2025-10-28 #8973 @pabl1t0x
Hi @ethanwhite. I have now addressed @dlebauer's comments, so I am regenerating the paper PDF. Thanks
