+++ reference/albedo_sat.md
@@ -1,0 +1,212 @@
+# Satellite albedo retrieval
+
+This function calculates narrowband and broadband albedo from surface
+reflectance data. The albedo retrieval method includes corrections for
+the anisotropic behavior of the reflected radiation field over snow and
+ice, and narrow-to-broadband albedo conversion algorithms.
+
+## Usage
+
+``` r
+albedo_sat(
+  SAA,
+  SZA,
+  VAA,
+  VZA,
+  slope,
+  aspect,
+  method,
+  blue = NULL,
+  green = NULL,
+  red = NULL,
+  NIR = NULL,
+  SWIR1 = NULL,
+  SWIR2 = NULL,
+  th
+)
+```
+
+## Arguments
+
+- SAA:
+
+  SpatRaster or numeric. Solar azimuth angle in degrees.
+
+- SZA:
+
+  SpatRaster or numeric. Solar zenith angle in degrees.
+
+- VAA:
+
+  SpatRaster or numeric. Satellite sensor azimuth angle in degrees.
+
+- VZA:
+
+  SpatRaster or numeric. Satellite sensor zenith angle in degrees.
+
+- slope:
+
+  SpatRaster or numeric. Surface slope in degrees.
+
+- aspect:
+
+  SpatRaster or numeric. Surface aspect in degrees.
+
+- method:
+
+  character. Number of spectral bands to calculate broadband albedo.
+  There are three options available: "twobands" (green and
+  near-infrared), "fourbands" (blue, green, red, near-infrared), and
+  "fivebands" (blue, red, near-infrared, shortwave-infrared 1, and
+  shortwave-infrared 2). If `method="twobands"` or `method="fivebands"`,
+  surface reflectance data is corrected for anisotropy before generating
+  broadband albedo. In contrast, `method="fourbands"` assumes Lambertian
+  reflection and converts surface reflectance to broadband albedo
+  directly. More details and examples about these workflows can be found
+  in the references below.
+
+- blue:
+
+  SpatRaster. Blue band surface reflectance (0.43-0.45 um). Required for
+  `method="fourbands"` and `method="fivebands"`.
+
+- green:
+
+  SpatRaster. Green band surface reflectance (0.53-0.59 um). Required
+  for all methods.
+
+- red:
+
+  SpatRaster. Red band surface reflectance (0.64-0.67 um). Required for
+  `method="fourbands"` and `method="fivebands"`.
+
+- NIR:
+
+  SpatRaster. Near-infrared surface reflectance (0.85-0.88 um). Required
+  for all methods.
+
+- SWIR1:
+
+  SpatRaster. Shortwave-infrared surface reflectance (1.57-1.65 um).
+  Required for `method="fivebands"`.
+
+- SWIR2:
+
+  SpatRaster. Shortwave-infrared surface reflectance (2.11-2.29 um).
+  Required for `method="fivebands"`.
+
+- th:
+
+  numeric. NDSII threshold to discriminate between snow and ice.
+
+## Value
+
+Returns a SpatRaster with four layers for `method="twobands"`, five
+layers for `method="fourbands"` and six layers for `method="fivebands"`.
+These layers are:
+
+- If `method="twobands"`: green and NIR narrowband albedo, broadband
+  albedo, and quality flags indicating whether broadband albedo was
+  calculated using the corrected green and NIR narrowband albedos
+  (`flag=1`) or the NIR albedo only (`flag=2`). Broadband albedo values
+  higher than one are not excluded for this method.
+
[…]
