+++ llms.txt
@@ -1,0 +1,187 @@
+# SatRbedo
+
+The `SatRbedo` package consists of a set of tools for retrieving snow
+and ice albedo from optical satellite imagery (e.g.,
+[Landsat](https://www.usgs.gov/landsat-missions/landsat-collection-2-surface-reflectance)
+and
+[Sentinel-2](https://dataspace.copernicus.eu/explore-data/data-collections/sentinel-data/sentinel-2)).
+`SatRbedo` functions require the following datasets:
+
+- Atmospherically corrected surface reflectance (gridded).
+- Satellite ($\varphi_{v}$, $\theta_{v}$) and solar ($\varphi_{s}$,
+  $\theta_{s}$) azimuth and zenith angles (numeric or gridded).
+- A digital elevation model (DEM) (gridded).
+- An outline of the area of interest (shapefile).
+
+The package includes tools for:
+
+- Image pre-processing: crop grids to a specified extent, project grids
+  with different coordinate systems, and convert data from integer to
+  floating point.
+- Convert nadir satellite observations to off-nadir values using
+  view-angle corrections.
+- Detect and mask topographic shadows.
+- Automatic discrimination of snow and ice surfaces.
+- Anisotropic correction of reflected radiation of glacier snow and ice
+  using empirical parameterizations of the bidirectional reflectance
+  distribution function (BRDF).
+- Topographic correction.
+- Narrow-to-broadband albedo conversion.
+
+## Installation
+
+You can install `SatRbedo` with:
+
+``` r
+install.packages('SatRbedo', repos = c('https://pabl1t0x.r-universe.dev', 'https://cloud.r-project.org'))
+```
+
+## Getting Started
+
+Snow and ice albedo retrieval from satellite imagery is a four-step
+workflow that includes image pre-processing, topographic and anisotropic
+corrections, and a narrow-to-broadband albedo conversion (Fig. 1).
+
+![Fig. 1 Flowchart of the satellite albedo retrieval
+workflow](reference/figures/workflow.png)
+
+Fig. 1 Flowchart of the satellite albedo retrieval workflow
+
+The following example shows how to estimate albedo using surface
+reflectance data from five spectral bands at the Athabasca Glacier in
+Canada on 11 September 2020.
+
+**Step 1: Load the data for the area of interest**
+
+``` r
+# Load the packages
+library(SatRbedo)
+library(terra)
+#> terra 1.8.70
+
+# Load the raw Sentinel-2 surface reflectance data
+# Note: each spectral band was previously cut out to the extent of the area of interest and renamed
+blue_SR <- system.file("extdata/athabasca_B02_20200911.tif", package = "SatRbedo") # blue band surface reflectance
+green_SR <- system.file("extdata/athabasca_B03_20200911.tif", package = "SatRbedo") # green band surface reflectance
+red_SR <- system.file("extdata/athabasca_B04_20200911.tif", package = "SatRbedo") # red band surface reflectance
+NIR_SR <- system.file("extdata/athabasca_B8A_20200911.tif", package = "SatRbedo") # near-infrared band surface reflectance
+SWIR1_SR <- system.file("extdata/athabasca_B11_20200911.tif", package = "SatRbedo") # shortwave-infrared band 1 surface reflectance
[…]
