+++ authors.md
@@ -1,0 +1,42 @@
+# Authors and Citation
+
+## Authors
+
+- **Pablo Fuchs**. Author, maintainer.
+  [](https://orcid.org/0000-0002-6042-5620)
+
+- **Ruzica Dadic**. Contributor, thesis advisor.
+  [](https://orcid.org/0000-0003-1303-1896)
+
+- **Shelley MacDonell**. Contributor, thesis advisor.
+  [](https://orcid.org/0000-0001-9641-4547)
+
+- **Heather Purdie**. Contributor, thesis advisor.
+  [](https://orcid.org/0000-0002-2723-6908)
+
+- **Brian Anderson**. Contributor, thesis advisor.
+
+- **Marwan Katurji**. Contributor, thesis advisor.
+  [](https://orcid.org/0000-0002-3368-1469)
+
+- **Javier Corripio**. Copyright holder.  
+  Author of the vectorial algebra algorithms used to detect topographic
+  shadows
+
+## Citation
+
+Source:
+[`inst/CITATION`](https://github.com/pabl1t0x/SatRbedo/blob/master/inst/CITATION)
+
+Fuchs P, Dadic R, MacDonell S, Purdie H, Anderson B, Katurji M (2025).
+*SatRbedo: An R package for retrieving snow and ice albedo from optical
+satellite imagery*. R package version 1.0,
+<https://github.com/pabl1t0x/SatRbedo>.
+
+    @Manual{,
+      title = {SatRbedo: An R package for retrieving snow and ice albedo from optical satellite imagery},
+      author = {Pablo Fuchs and Ruzica Dadic and Shelley MacDonell and Heather Purdie and Brian Anderson and Marwan Katurji},
+      year = {2025},
+      note = {R package version 1.0},
+      url = {https://github.com/pabl1t0x/SatRbedo},
+    }
