+++ reference/index.md
@@ -1,31 +1,51 @@
 # Package index
 
-## All functions
+## Preprocessing
 
+Functions for satellite data preprocessing
+
+- [`cfactor_BRDF()`](https://pabl1t0x.github.io/SatRbedo/reference/cfactor_BRDF.md)
+  : c-factor BRDF normalization
+- [`preproc()`](https://pabl1t0x.github.io/SatRbedo/reference/preproc.md)
+  : Satellite data pre-processing
+
+## Topographic correction
+
+Functions for correcting surface reflectance for the effects of
+topography
+
+- [`cast_shadows()`](https://pabl1t0x.github.io/SatRbedo/reference/cast_shadows.md)
+  : Cast shadows
+- [`hill_shade()`](https://pabl1t0x.github.io/SatRbedo/reference/hill_shade.md)
+  : Topographic hillshading
+- [`shadow_removal()`](https://pabl1t0x.github.io/SatRbedo/reference/shadow_removal.md)
+  : Topographic shadows
+- [`topo_corr()`](https://pabl1t0x.github.io/SatRbedo/reference/topo_corr.md)
+  : Topographic correction
+- [`topo_splot()`](https://pabl1t0x.github.io/SatRbedo/reference/topo_splot.md)
+  : Reflectance-IC scatterplot
+
+## Anisotropic correction
+
+Functions for calculating the anisotropic reflection factors for
+different spectral bands and surface types (snow or ice)
+
+- [`f_BRDF()`](https://pabl1t0x.github.io/SatRbedo/reference/f_BRDF.md)
+  : Anisotropy correction of reflected radiation over melting snow and
+  glacier ice
 - [`NDSII_hist()`](https://pabl1t0x.github.io/SatRbedo/reference/NDSII_hist.md)
   : NDSII histogram
+- [`snow_or_ice()`](https://pabl1t0x.github.io/SatRbedo/reference/snow_or_ice.md)
+  : Automatic discrimination of snow and ice cover types
+
+## Narrow-to-broadband albedo conversion
+
+Functions for calculating narrowband and broadband albedo from
+topographically-corrected surface reflectance
+
 - [`albedo_Knap()`](https://pabl1t0x.github.io/SatRbedo/reference/albedo_broad.md)
   [`albedo_Liang()`](https://pabl1t0x.github.io/SatRbedo/reference/albedo_broad.md)
   [`albedo_Feng()`](https://pabl1t0x.github.io/SatRbedo/reference/albedo_broad.md)
   : Narrow-to-broadband albedo conversion
 - [`albedo_sat()`](https://pabl1t0x.github.io/SatRbedo/reference/albedo_sat.md)
   : Satellite albedo retrieval
-- [`cast_shadows()`](https://pabl1t0x.github.io/SatRbedo/reference/cast_shadows.md)
-  : Cast shadows
-- [`cfactor_BRDF()`](https://pabl1t0x.github.io/SatRbedo/reference/cfactor_BRDF.md)
-  : c-factor BRDF normalization
-- [`f_BRDF()`](https://pabl1t0x.github.io/SatRbedo/reference/f_BRDF.md)
-  : Anisotropy correction of reflected radiation over melting snow and
-  glacier ice
-- [`hill_shade()`](https://pabl1t0x.github.io/SatRbedo/reference/hill_shade.md)
-  : Topographic hillshading
-- [`preproc()`](https://pabl1t0x.github.io/SatRbedo/reference/preproc.md)
-  : Satellite data pre-processing
-- [`shadow_removal()`](https://pabl1t0x.github.io/SatRbedo/reference/shadow_removal.md)
-  : Topographic shadows
[…]
