### 2025-11-18 e2540f6
Add NEWS.md
