+++ NEWS.md
@@ -1,0 +1,3 @@
+# SatRbedo 1.0
+
+* Initial release.
