+++ news/index.md
@@ -1,0 +1,5 @@
+# Changelog
+
+## SatRbedo 1.0
+
+- Initial release.
