+++ DESCRIPTION
@@ -75,8 +75,11 @@ Remotes:
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.3
 Suggests: 
+    knitr,
+    rmarkdown,
     testthat (>= 3.0.0),
     vdiffr
 Config/testthat/edition: 3
 URL: https://github.com/pabl1t0x/SatRbedo, https://pabl1t0x.github.io/SatRbedo/
 BugReports: https://github.com/pabl1t0x/SatRbedo/issues
+VignetteBuilder: knitr

[…]
