+++ articles/index.md
@@ -1,0 +1,5 @@
+# Articles
+
+### All vignettes
+
+- [SatRbedo](https://pabl1t0x.github.io/SatRbedo/articles/SatRbedo.md):
