+++ README.md
@@ -12,12 +12,19 @@ coverage](https://codecov.io/gh/pabl1t0x/SatRbedo/branch/master/graph/badge.svg)
 [![status](https://joss.theoj.org/papers/c5899414306d4cf318350654b7aa38fa/status.svg)](https://joss.theoj.org/papers/c5899414306d4cf318350654b7aa38fa)
 <!-- badges: end -->
 
-The `SatRbedo` package consists of a set of tools for retrieving snow
-and ice albedo from optical satellite imagery (e.g.,
+Albedo is a key parameter that controls the energy exchange between the
+atmosphere and snow or ice surfaces, playing a crucial role in
+estimating glacier melt and mass balance. While commonly measured in the
+field with pyranometers, these instruments have limited spatial coverage
+due to their footprint. Therefore, remote sensing offers the best option
+for deriving snow and ice albedo data and analysing its spatial and
+temporal variability. `SatRbedo` is an extensible, well-documented
+package for retrieving snow and ice albedo from optical satellite
+images, such as
 [Landsat](https://www.usgs.gov/landsat-missions/landsat-collection-2-surface-reflectance)
 and
-[Sentinel-2](https://dataspace.copernicus.eu/explore-data/data-collections/sentinel-data/sentinel-2)).
-`SatRbedo` functions require the following datasets:
+[Sentinel-2](https://dataspace.copernicus.eu/explore-data/data-collections/sentinel-data/sentinel-2).
+It requires the following input datasets:
 
 - Atmospherically corrected surface reflectance (gridded).
 - Satellite ($\varphi_v$, $\theta_v$) and solar ($\varphi_s$,
@@ -25,35 +32,32 @@ and
 - A digital elevation model (DEM) (gridded).
 - An outline of the area of interest (shapefile).
 
-The package includes the following functions:
+The `SatRbedo` package has been designed for researchers, consultants,
+and individuals interested in exploring changes in snow and ice albedo.
+It includes the following functions:
 
-- `preproc()`: Image pre-processing, including crop grids to a specified
-  extent, project grids with different coordinate systems, and convert
-  data from integer to floating point.
-
-- `cfactor_BRDF`: Convert nadir satellite observations to off-nadir
+- `preproc()` provides tools for image pre-processing, including
+  cropping grids to a specified extent, projecting grids with different
+  coordinate systems, and converting data from integer to floating
+  point.
+- `cfactor_BRDF` converts nadir satellite observations to off-nadir
   values using view-angle corrections.
-
-- `cast_shadows()`, `hill_shade()`, and `shadow_removal()`: Detect and
+- `cast_shadows()`, `hill_shade()`, and `shadow_removal()` detect and
   mask topographic shadows.
-
-- `topo_corr()`, and `topo_splot()`: Topographic correction.
-
-- `snow_or_ice()`, and `NDSII_hist()`: Automatic discrimination of snow
-  and ice surfaces.
-
-- `f_BRDF()`: Anisotropic correction of reflected radiation of glacier
-  snow and ice using empirical parameterizations of the bidirectional
-  reflectance distribution function (BRDF).
-
-- `albedo_sat()`: This function calculates narrowband and broadband
-  albedo from surface reflectance data. The albedo retrieval methods
-  include corrections for the anisotropic behavior of the reflected
-  radiation field over snow and ice, and narrow-to-broadband albedo
-  conversion algorithms.
-
-- `albedo_Knap()`, `albedo_Feng()`, and `albedo_Liang()`:
-  Narrow-to-broadband albedo conversion.
+- `topo_corr()` and `topo_splot()` provide tools to correct surface
+  reflectance for the effects of topography and to visualise the
[…]
