### 2025-11-27 0b8d4bc
Updated vignettes
