+++ vignettes/albedo_retrieval_landsat.Rmd
@@ -1,0 +1,19 @@
+---
+title: "albedo_retrieval_landsat"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{albedo_retrieval_landsat}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup}
+library(SatRbedo)
+```
