+++ vignettes/SatRbedo.Rmd
@@ -1,8 +1,8 @@
 ---
-title: "SatRbedo"
+title: "Introduction to SatRbedo"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{SatRbedo}
+  %\VignetteIndexEntry{Introduction to SatRbedo}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -14,15 +14,18 @@ knitr::opts_chunk$set(
 )
 ```
 
-```{r setup}
+This document aims to introduce users to the basic functionality of `SatRbedo`. The example demonstrates how to use the package to compute snow and ice albedo at the Athabasca Glacier in Canada from five spectral bands. It utilises Sentinel-2 surface reflectance data measured on 11 September 2020.
+
+First, we load the `satRbedo` and `{terra}` packages. The latter is used for spatial data manipulation and analysis.
+
+```{r setup, message = FALSE}
 library(SatRbedo)
 library(terra)
 ```
 
- This document aims to introduce the users to the basic functionality of `SatRbedo`. Here we use
- Sentinel-2 data from five spectral bands at the Athabasca Glacier in Canada on 11 September 2020.
+The process for generating satellite albedo retrievals involves four steps:
 
- **Step 1: Load the data for the area of interest**
+## Step 1: Load the data for the area of interest
 
 <!-- par(mar=c(1,1,1,1)) -->
 
@@ -42,7 +45,7 @@ dem <- system.file("extdata/athabasca_dem.tif", package = "SatRbedo")
 outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
 ```
 
-**Step 2: Data pre-processing**
+## Step 2: Data pre-processing
 
 ```{r example-preprocessing, tidy = 'styler'}
 # Transform the input data to SpatRaster and crop to the area of interest
@@ -55,7 +58,7 @@ swir1 <- preproc(grd = SWIR1_SR, outline = outline)
 swir2 <- preproc(grd = SWIR2_SR, outline = outline)
 ```
 
-**Step 3: Topographic correction**
+## Step 3: Topographic correction
 
 ```{r example-topography, tidy = 'styler'}
 SAA <- 164.8 # solar azimuth angle
@@ -68,9 +71,9 @@ swir1_corr <- topo_corr(swir1, dem, SAA, SZA)
 swir2_corr <- topo_corr(swir2, dem, SAA, SZA)
 ```
 
-**Step 4: Estimation of broadband albedo after anisotropic correction**
+## Step 4: Estimation of broadband albedo after anisotropic correction
 
-```{r example-anisotropy, tidy = 'styler', fig.dim = c(5, 5), fig.cap = "Broadband albedo"}
+```{r example-anisotropy, tidy = 'styler', fig.dim = c(6, 6)}
 SAA <- 164.8 # solar azimuth angle
 SZA <- 48.9 # solar zenith angle
 VAA <- 90.9 # view azimuth angle
@@ -86,10 +89,9 @@ broadband_albedo <- albedo_sat(
   th = threshold
 )
 # Plot the results
-plot(broadband_albedo[[6]])
+plot(broadband_albedo[[6]], plg = list(title = "Albedo"))
 ```
 
-# Where to go next?
+## Where to go next?
 
-You can check function documentation here and a more elaborate example using Landsat data here.
-
+You can check the function documentation [here](https://pabl1t0x.github.io/SatRbedo/reference/index.html). A more elaborate example using Landsat data can be found in `vignette("albedo_retrieval_landsat")`.
