+++ articles/albedo_retrieval_landsat.md
@@ -1,0 +1,5 @@
+# albedo_retrieval_landsat
+
+``` r
+library(SatRbedo)
+```
