+++ llms.txt
@@ -167,4 +167,6 @@ topographically-corrected surface reflectance
 
 ### All vignettes
 
-- [SatRbedo](https://pabl1t0x.github.io/SatRbedo/articles/SatRbedo.md):
+- [albedo_retrieval_landsat](https://pabl1t0x.github.io/SatRbedo/articles/albedo_retrieval_landsat.md):
+- [Introduction to
+  SatRbedo](https://pabl1t0x.github.io/SatRbedo/articles/SatRbedo.md):
