+++ articles/index.md
@@ -2,4 +2,6 @@
 
 ### All vignettes
 
-- [SatRbedo](https://pabl1t0x.github.io/SatRbedo/articles/SatRbedo.md):
+- [albedo_retrieval_landsat](https://pabl1t0x.github.io/SatRbedo/articles/albedo_retrieval_landsat.md):
+- [Introduction to
+  SatRbedo](https://pabl1t0x.github.io/SatRbedo/articles/SatRbedo.md):
