+++ pkgdown.yml
@@ -2,8 +2,9 @@ pandoc: 3.1.11
 pkgdown: 2.2.0
 pkgdown_sha: ~
 articles:
+  albedo_retrieval_landsat: albedo_retrieval_landsat.html
   SatRbedo: SatRbedo.html
-last_built: 2025-11-26T21:37Z
+last_built: 2025-11-27T14:42Z
 urls:
   reference: https://pabl1t0x.github.io/SatRbedo/reference
   article: https://pabl1t0x.github.io/SatRbedo/articles
