+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   albedo_retrieval_landsat: albedo_retrieval_landsat.html
   SatRbedo: SatRbedo.html
-last_built: 2025-11-27T14:42Z
+last_built: 2025-11-27T14:45Z
 urls:
   reference: https://pabl1t0x.github.io/SatRbedo/reference
   article: https://pabl1t0x.github.io/SatRbedo/articles
