+++ R/snow_or_ice.R
@@ -36,8 +36,8 @@
 #' @seealso [preproc()], [EBImage::otsu()], [terra::plot()]
 #'
 #' @examples
-#' green <- system.file("extdata/athabasca_B03_20200911.tif", package = "SatRbedo")
-#' nir <- system.file("extdata/athabasca_B8A_20200911.tif", package = "SatRbedo")
+#' green <- system.file("extdata/athabasca_2020253_B03_S30.tif", package = "SatRbedo")
+#' nir <- system.file("extdata/athabasca_2020253_B8A_S30.tif", package = "SatRbedo")
 #' outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
 #' green <- preproc(grd = green, outline = outline)
 #' nir <- preproc(grd = nir, outline = outline)
@@ -74,8 +74,8 @@ snow_or_ice <- function(green, NIR) {
 #' @seealso [preproc()], [snow_or_ice()], [graphics::hist()]
 #'
 #' @examples
-#' green <- system.file("extdata/athabasca_B03_20200911.tif", package = "SatRbedo")
-#' nir <- system.file("extdata/athabasca_B8A_20200911.tif", package = "SatRbedo")
+#' green <- system.file("extdata/athabasca_2020253_B03_S30.tif", package = "SatRbedo")
+#' nir <- system.file("extdata/athabasca_2020253_B8A_S30.tif", package = "SatRbedo")
 #' outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
 #' green <- preproc(grd = green, outline = outline)
 #' nir <- preproc(grd = nir, outline = outline)
