+++ R/topo_corr.R
@@ -49,10 +49,10 @@
 #' @examples
 #' library(terra)
 #' outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
-#' nir <- system.file("extdata/athabasca_B8A_20200911.tif", package = "SatRbedo")
+#' nir <- system.file("extdata/athabasca_2020253_B8A_S30.tif", package = "SatRbedo")
 #' dem <- system.file("extdata/athabasca_dem.tif", package = "SatRbedo")
-#' nir <- preproc(grd = nir, outline = outline)
-#' dem <- preproc(grd = dem, outline = outline)
+#' nir <- preproc(grd = nir)
+#' dem <- preproc(grd = dem)
 #' SAA <- 164.8
 #' SZA <- 48.9
 #'
@@ -112,10 +112,10 @@ topo_corr <- function(band, dem, SAA, SZA, method = "tanrotation", IC_min = NULL
 #'
 #' @examples
 #' outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
-#' nir <- system.file("extdata/athabasca_B8A_20200911.tif", package = "SatRbedo")
+#' nir <- system.file("extdata/athabasca_2020253_B8A_S30.tif", package = "SatRbedo")
 #' dem <- system.file("extdata/athabasca_dem.tif", package = "SatRbedo")
-#' nir <- preproc(grd = nir, outline = outline)
-#' dem <- preproc(grd = dem, outline = outline)
+#' nir <- preproc(grd = nir)
+#' dem <- preproc(grd = dem)
 #' SAA <- 164.8
 #' SZA <- 48.9
 #' corr <- topo_corr(band = nir, dem = dem, SAA = SAA, SZA = SZA, method = "tanrotation")
