+++ articles/SatRbedo.md
@@ -4,7 +4,7 @@ This document aims to introduce users to the basic functionality of
 `SatRbedo`. The example demonstrates how to use the package to compute
 snow and ice albedo at the Athabasca Glacier in Canada from five
 spectral bands. It utilises Sentinel-2 surface reflectance data measured
-on 11 September 2020.
+on 9 September 2020.
 
 First, we load the `satRbedo` and [terra](https://rspatial.org/)
 packages. The latter is used for spatial data manipulation and analysis.
@@ -14,7 +14,7 @@ library(SatRbedo)
 library(terra)
 ```
 
-The process for generating satellite albedo retrievals involves four
+The process for generating satellite albedo retrievals involves five
 steps:
 
 ## Step 1: Load the data for the area of interest
@@ -22,12 +22,12 @@ steps:
 ``` r
 # Load the raw Sentinel-2 surface reflectance data
 # Note: each spectral band was previously cut out to the extent of the area of interest and renamed
-blue_SR <- system.file("extdata/athabasca_B02_20200911.tif", package = "SatRbedo") # blue band surface reflectance
-green_SR <- system.file("extdata/athabasca_B03_20200911.tif", package = "SatRbedo") # green band surface reflectance
-red_SR <- system.file("extdata/athabasca_B04_20200911.tif", package = "SatRbedo") # red band surface reflectance
-NIR_SR <- system.file("extdata/athabasca_B8A_20200911.tif", package = "SatRbedo") # near-infrared band surface reflectance
-SWIR1_SR <- system.file("extdata/athabasca_B11_20200911.tif", package = "SatRbedo") # shortwave-infrared band 1 surface reflectance
-SWIR2_SR <- system.file("extdata/athabasca_B12_20200911.tif", package = "SatRbedo") # shortwave-infrared band 2 surface reflectance
+blue_SR <- system.file("extdata/athabasca_2020253_B02_S30.tif", package = "SatRbedo") # blue band surface reflectance
+green_SR <- system.file("extdata/athabasca_2020253_B03_S30.tif", package = "SatRbedo") # green band surface reflectance
+red_SR <- system.file("extdata/athabasca_2020253_B04_S30.tif", package = "SatRbedo") # red band surface reflectance
+NIR_SR <- system.file("extdata/athabasca_2020253_B8A_S30.tif", package = "SatRbedo") # near-infrared band surface reflectance
+SWIR1_SR <- system.file("extdata/athabasca_2020253_B11_S30.tif", package = "SatRbedo") # shortwave-infrared band 1 surface reflectance
+SWIR2_SR <- system.file("extdata/athabasca_2020253_B12_S30.tif", package = "SatRbedo") # shortwave-infrared band 2 surface reflectance
 
 # Load the DEM and the outline
 # Note: the DEM was re-projected to the extent of the area of interest
@@ -38,46 +38,81 @@ outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
 ## Step 2: Data pre-processing
 
 ``` r
-# Transform the input data to SpatRaster and crop to the area of interest
+# Transform the input data to SpatRaster and crop to the extent of the area of interest
 dem <- terra::rast(dem)
-blue <- preproc(grd = blue_SR, outline = outline)
-green <- preproc(grd = green_SR, outline = outline)
[…]
