### 2025-12-01 43fa173
Updated function documentation
