+++ reference/f_BRDF.md
@@ -136,8 +136,6 @@ f <- f_BRDF(
   method = "twobands", green = green, NIR = nir,
   th = th
 )
-#> Warning: [mask] CRS do not match
-#> Warning: [mask] CRS do not match
 plot(f[[1]])
 
 plot(f[[2]])
@@ -151,8 +149,6 @@ f <- f_BRDF(
   method = "fivebands", green = green, NIR = nir,
   th = th
 )
-#> Warning: [mask] CRS do not match
-#> Warning: [mask] CRS do not match
 plot(f[[1]])
 
 plot(f[[2]])
