### 2025-12-05 e81d137
Updated JOSS paper
