+++ inst/paper/paper.md
@@ -32,7 +32,7 @@ affiliations:
    index: 2
  - name: Antarctic Research Centre, Victoria University, Wellington, New Zealand
    index: 3
-date: 12 June 2025
+date: 5 December 2025
 bibliography: paper.bib
 ---
 
@@ -42,32 +42,18 @@ Albedo is a key variable determining the amount of solar radiation absorbed by
 snow and ice surfaces. As such, it influences meltwater production, glacier mass
 balance, and the energy exchange between the Earth and the atmosphere [@jonsell_spatial_2003;@hock_glacier_2005]. Satellite remote sensing has been widely recognized as the best practical
 approach for monitoring and mapping surface albedo across different spatial and temporal scales [@lin_estimating_2022;@urraca_assessing_2023]. Here, we present the `SatRbedo` R package:
-an extensible, standalone toolbox for retrieving snow and ice albedo from optical satellite imagery.
-The package includes functions for image preprocessing, converting nadir satellite observations to off-nadir values using view-angle corrections, detecting topographic shadows, discriminating snow and ice surfaces, correcting for topographic effects and the anisotropic behavior of reflected radiation of glacier snow and ice, and converting narrowband to broadband albedo. The package has a modular structure that allows for changing the implemented routines and provides output that can be used independently
-or as input to other functions. `SatRbedo` is designed to work with medium-resolution satellite data
-(e.g., Landsat and Sentinel-2), although data from different satellite sensors can also be used.
+an extensible, standalone toolbox for retrieving snow and ice albedo from medium-resolution multispectral optical satellite imagery. The package includes functions for image preprocessing, converting nadir satellite observations to off-nadir values using view-angle corrections, detecting topographic shadows, discriminating snow and ice surfaces, correcting for topographic effects and the anisotropic behavior of radiation reflected by glacier snow and ice, and converting narrowband to broadband albedo. The package has a modular structure that allows for changing the implemented routines and provides output that can be used independently or as input to other functions. `SatRbedo` is optimized for Landsat and Sentinel-2 data but can also process data from other medium-resolution sensors (e.g., ALOS/AVNIR-2, SPOT, and ASTER), provided they meet the required input data specifications. For example, if at-sensor radiance data is available, atmospheric correction is required before using the package. Additionally, cross-sensor calibration should be performed to minimize spectral differences between sensors.
 
 # Statement of need
 
-The land surface albedo is an essential climate variable that controls the partitioning of radiative energy between the surface and the atmosphere [@bojinski_concept_2014;@radeloff_need_2024].
-Albedo is the hemispherically integrated reflectance representing the proportion of the incoming solar
-radiation reflected from a unit surface area [@budyko_effect_1969;@schaepman-strub_reflectance_2006]. In
-the cryosphere, albedo ranges from <0.1 for debris-covered ice to 0.3-0.4 for bare ice to ~0.5 for aged, wet snow to >0.9 for fresh, dry snow [@cuffey_physics_2010].
[…]
