+++ inst/paper/paper.bib
@@ -32,17 +32,17 @@
 	pages = {1431--1443},
 }
 
-@article{budyko_effect_1969,
-	title = {The effect of solar radiation variations on the climate of the {Earth}},
-	volume = {21},
-	doi = {10.1111/j.2153-3490.1969.tb00466.x},
-	number = {5},
-	journal = {Tellus},
-	author = {Budyko, M. I.},
-	month = oct,
-	year = {1969},
-	pages = {611--619},
-}
+% @article{budyko_effect_1969,
+% 	title = {The effect of solar radiation variations on the climate of the {Earth}},
+% 	volume = {21},
+% 	doi = {10.1111/j.2153-3490.1969.tb00466.x},
+% 	number = {5},
+% 	journal = {Tellus},
+% 	author = {Budyko, M. I.},
+% 	month = oct,
+% 	year = {1969},
+% 	pages = {611--619},
+% }
 
 @article{carlsen_parameterizing_2020,
 	title = {Parameterizing anisotropic reflectance of snow surfaces from airborne digital camera observations in {Antarctica}},
@@ -108,16 +108,16 @@
 	pages = {1--29},
 }
 
-@article{flanner_snicar-adv3_2021,
-	title = {{SNICAR}-{ADv3}: a community tool for modeling spectral snow albedo},
-	volume = {14},
-	doi = {10.5194/gmd-14-7673-2021},
-	number = {12},
-	journal = {Geoscientific Model Development},
-	author = {Flanner, Mark G. and Arnheim, Julian B. and Cook, Joseph M. and Dang, Cheng and He, Cenlin and Huang, Xianglei and Singh, Deepak and Skiles, S. McKenzie and Whicker, Chloe A. and Zender, Charles S.},
-	year = {2021},
-	pages = {7673--7704},
-}
+% @article{flanner_snicar-adv3_2021,
+% 	title = {{SNICAR}-{ADv3}: a community tool for modeling spectral snow albedo},
+% 	volume = {14},
+% 	doi = {10.5194/gmd-14-7673-2021},
+% 	number = {12},
+% 	journal = {Geoscientific Model Development},
+% 	author = {Flanner, Mark G. and Arnheim, Julian B. and Cook, Joseph M. and Dang, Cheng and He, Cenlin and Huang, Xianglei and Singh, Deepak and Skiles, S. McKenzie and Whicker, Chloe A. and Zender, Charles S.},
+% 	year = {2021},
+% 	pages = {7673--7704},
+% }
 
 @article{fugazza_spatial_2016,
 	title = {Spatial distribution of surface albedo at the {Forni} {Glacier} ({Stelvio} {National} {Park}, {Central} {Italian} {Alps})},
@@ -129,17 +129,17 @@
 	pages = {128--137},
 }
 
-@article{gaillard_improving_2025,
-	title = {Improving large-scale snow albedo modeling using a climatology of light-absorbing particle deposition},
-	volume = {19},
-	doi = {10.5194/tc-19-769-2025},
-	number = {2},
-	journal = {The Cryosphere},
-	author = {Gaillard, Manon and Vionnet, Vincent and Lafaysse, Matthieu and Dumont, Marie and Ginoux, Paul},
-	month = feb,
-	year = {2025},
-	pages = {769--792},
-}
+% @article{gaillard_improving_2025,
+% 	title = {Improving large-scale snow albedo modeling using a climatology of light-absorbing particle deposition},
+% 	volume = {19},
+% 	doi = {10.5194/tc-19-769-2025},
+% 	number = {2},
+% 	journal = {The Cryosphere},
+% 	author = {Gaillard, Manon and Vionnet, Vincent and Lafaysse, Matthieu and Dumont, Marie and Ginoux, Paul},
[…]
