+++ inst/paper/paper.md
@@ -48,7 +48,7 @@ an extensible, standalone toolbox for retrieving snow and ice albedo from medium
 
 The land surface albedo is an essential climate variable that controls the partitioning of radiative energy between the surface and the atmosphere [@bojinski_concept_2014;@radeloff_need_2024]. In the cryosphere, albedo ranges from <0.1 for debris-covered ice to 0.3-0.4 for bare ice to ~0.5 for aged, wet snow to >0.9 for fresh, dry snow [@cuffey_physics_2010]. Snow and ice albedo depend on the inherent optical properties of the surface (including snow grain size and shape, snowpack thickness, surface roughness, and water and impurity content) and are also influenced by environmental conditions (apparent albedo), including the angular and spectral distribution of solar radiation, topography, the underlying substrate for thin snow cover, and cloud cover [@warren_optical_2019;@whicker_snicar-adv4_2022].
 
-Albedo can be measured in the field using observations from a pyranometer pair, one looking upward and the other looking downward [@driemel_baseline_2018;@picard_spectral_2020]. Although these instruments provide highly accurate point measurements, they have limited spatial coverage due to their relatively small footprint. Alternatively, albedo can be estimated from satellite remote sensing[@bertoncini_large-area_2022;@fugazza_spatial_2016], which offers the best option for studying albedo changes, considering its high spatial and temporal variability [@berg_finescale_2020].
+Albedo can be measured in the field using observations from a pyranometer pair, one looking upward and the other looking downward [@driemel_baseline_2018;@picard_spectral_2020]. Although these instruments provide highly accurate point measurements, they have limited spatial coverage due to their relatively small footprint. Alternatively, albedo can be estimated from satellite remote sensing [@bertoncini_large-area_2022;@fugazza_spatial_2016], which offers the best option for studying albedo changes, considering its high spatial and temporal variability [@berg_finescale_2020].
 
 Satellite albedo retrievals typically comprise three steps: atmospheric correction, modeling of the angular reflectance, and narrow-to-broadband albedo conversion [@carlsen_parameterizing_2020;@qu_mapping_2015]. The algorithms for atmospheric correction [@doxani_atmospheric_2023;@vermote_preliminary_2016], angular reflectance modelling [@klok_temporal_2003;@lucht_algorithm_2000;@ren_anisotropy_2021], and narrow-to-broadband albedo conversion
 [@knap_narrowband_1999;@li_preliminary_2018;@liang_narrowband_2001] are well established and validated across a number of case studies. In addition to these steps, satellite image preprocessing and topographic correction are necessary to homogenize the input data and minimize the effects of slope and aspect on albedo, respectively.
