### 2025-12-07 945007b
Updated vignettes
