### 2025-12-07 ddd940d
Updated vignettes
