### 2025-12-08 e264f16
Updated vignettes
