+++ vignettes/albedo_retrieval_landsat.Rmd
@@ -31,6 +31,128 @@ Landsat surface reflectance (Landsat Collection 2 Level-2) and topographic (e.g.
 
 The `SatRbedo` package provides functions to calculate albedo using two (green and near-infrared), four (blue, green, red, and near-infrared), and five (blue, red, near-infrared, shortwave-infrared 1, and shortwave-infrared 2) spectral bands. These functions also require the solar and view angles. To generate these angles, the [Landsat Angles Creation Tools](https://www.usgs.gov/landsat-missions/solar-illumination-and-sensor-viewing-angle-coefficient-files) can be used. 
 
-After downloading the necessary input data, reproject the digital elevation model (DEM) grid and the glacier outline shapefile to Landsat’s coordinate system, then resample them to 30 m spatial resolution. Crop the DEM, spectral surface reflectance, and angle bands to the area of interest. The tasks can be performed using the `preproc()` function and the `{terra}` package.
+After downloading the necessary input data, reproject the digital elevation model (DEM) grid and the glacier outline shapefile to Landsat’s coordinate system, then resample the DEM to 30 m spatial resolution. Select scenes with minimal cloud coverage and crop the DEM, surface reflectance, and angle bands to the area of interest. These tasks can be performed using the `preproc()` function and the `{terra}` package.
 
-This workflow can be applied tor other instruments and sensors, by specifying the spectral bands in the functions.
+Below, we provide the necessary input data to run the `SatRbedo` functions for the area of interest:
+
+```{r input-data, tidy = 'styler'}
+# Load the Landsat surface reflectance bands
+blue_SR <- system.file("extdata/athabasca_2020229_B02_L30.tif", package = "SatRbedo") # blue band surface reflectance (Landsat band 2)
+green_SR <- system.file("extdata/athabasca_2020229_B03_L30.tif", package = "SatRbedo") # green band surface reflectance (Landsat band 3)
+red_SR <- system.file("extdata/athabasca_2020229_B04_L30.tif", package = "SatRbedo") # red band surface reflectance (Landsat band 4)
+NIR_SR <- system.file("extdata/athabasca_2020229_B05_L30.tif", package = "SatRbedo") # near-infrared band surface reflectance (Landsat band 5)
+SWIR1_SR <- system.file("extdata/athabasca_2020229_B06_L30.tif", package = "SatRbedo") # shortwave-infrared band 1 surface reflectance (Landsat band 6)
+SWIR2_SR <- system.file("extdata/athabasca_2020229_B07_L30.tif", package = "SatRbedo") # shortwave-infrared band 2 surface reflectance (Landsat band 7)
+
+# Load the DEM and the glacier outline
+dem <- system.file("extdata/athabasca_dem.tif", package = "SatRbedo")
+outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
+
+# Transform the raster data to SpatRaster and the glacier outline to SpatVector
+dem <- terra::rast(dem)
+glacier_mask <- terra::vect(outline)
+blue <- terra::rast(blue_SR)
+green <- terra::rast(green_SR)
+red <- terra::rast(red_SR)
+nir <- terra::rast(NIR_SR)
+swir1 <- terra::rast(SWIR1_SR)
+swir2 <- terra::rast(SWIR2_SR)
+```
+
+## Topographic correction
+
[…]
