+++ articles/albedo_retrieval_landsat.md
@@ -37,11 +37,152 @@ can be used.
 
 After downloading the necessary input data, reproject the digital
 elevation model (DEM) grid and the glacier outline shapefile to
-Landsat’s coordinate system, then resample them to 30 m spatial
-resolution. Crop the DEM, spectral surface reflectance, and angle bands
-to the area of interest. The tasks can be performed using the
+Landsat’s coordinate system, then resample the DEM to 30 m spatial
+resolution. Select scenes with minimal cloud coverage and crop the DEM,
+surface reflectance, and angle bands to the area of interest. These
+tasks can be performed using the
 [`preproc()`](https://pabl1t0x.github.io/SatRbedo/reference/preproc.md)
 function and the [terra](https://rspatial.org/) package.
 
-This workflow can be applied tor other instruments and sensors, by
-specifying the spectral bands in the functions.
+Below, we provide the necessary input data to run the `SatRbedo`
+functions for the area of interest:
+
+``` r
+# Load the Landsat surface reflectance bands
+blue_SR <- system.file("extdata/athabasca_2020229_B02_L30.tif", package = "SatRbedo") # blue band surface reflectance (Landsat band 2)
+green_SR <- system.file("extdata/athabasca_2020229_B03_L30.tif", package = "SatRbedo") # green band surface reflectance (Landsat band 3)
+red_SR <- system.file("extdata/athabasca_2020229_B04_L30.tif", package = "SatRbedo") # red band surface reflectance (Landsat band 4)
+NIR_SR <- system.file("extdata/athabasca_2020229_B05_L30.tif", package = "SatRbedo") # near-infrared band surface reflectance (Landsat band 5)
+SWIR1_SR <- system.file("extdata/athabasca_2020229_B06_L30.tif", package = "SatRbedo") # shortwave-infrared band 1 surface reflectance (Landsat band 6)
+SWIR2_SR <- system.file("extdata/athabasca_2020229_B07_L30.tif", package = "SatRbedo") # shortwave-infrared band 2 surface reflectance (Landsat band 7)
+
+# Load the DEM and the glacier outline
+dem <- system.file("extdata/athabasca_dem.tif", package = "SatRbedo")
+outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
+
+# Transform the raster data to SpatRaster and the glacier outline to SpatVector
+dem <- terra::rast(dem)
+glacier_mask <- terra::vect(outline)
+blue <- terra::rast(blue_SR)
+green <- terra::rast(green_SR)
+red <- terra::rast(red_SR)
+nir <- terra::rast(NIR_SR)
+swir1 <- terra::rast(SWIR1_SR)
+swir2 <- terra::rast(SWIR2_SR)
+```
+
+## Topographic correction
+
+The topographic correction compensates for differences in solar
+illumination caused by elevation, slope, aspect, and the obstruction of
+surrounding terrain. The topographic correction methods implemented in
+`SatRbedo` require the removal of terrain shadow effects, including self
+and cast shadows. This can be achieved with the
+[`shadow_removal()`](https://pabl1t0x.github.io/SatRbedo/reference/shadow_removal.md)
+function:
+
+``` r
+SAA <- 154.6 # solar azimuth angle
+SZA <- 40.8 # solar zenith angle
[…]
