+++ articles/SatRbedo.md
@@ -7,7 +7,8 @@ spectral bands. It utilises Sentinel-2 surface reflectance data measured
 on 9 September 2020.
 
 First, we load the `SatRbedo` and [terra](https://rspatial.org/)
-packages. The latter is used for spatial data manipulation and analysis.
+packages. The latter is used for spatial data manipulation,
+visualization, and analysis.
 
 ``` r
 library(SatRbedo)
@@ -20,17 +21,18 @@ steps:
 ## Step 1: Load the data for the area of interest
 
 ``` r
-# Load the raw Sentinel-2 surface reflectance data
+# Load Sentinel-2 surface reflectance data
 # Note: each spectral band was previously cut out to the extent of the area of interest and renamed
-blue_SR <- system.file("extdata/athabasca_2020253_B02_S30.tif", package = "SatRbedo") # blue band surface reflectance
-green_SR <- system.file("extdata/athabasca_2020253_B03_S30.tif", package = "SatRbedo") # green band surface reflectance
-red_SR <- system.file("extdata/athabasca_2020253_B04_S30.tif", package = "SatRbedo") # red band surface reflectance
-NIR_SR <- system.file("extdata/athabasca_2020253_B8A_S30.tif", package = "SatRbedo") # near-infrared band surface reflectance
-SWIR1_SR <- system.file("extdata/athabasca_2020253_B11_S30.tif", package = "SatRbedo") # shortwave-infrared band 1 surface reflectance
-SWIR2_SR <- system.file("extdata/athabasca_2020253_B12_S30.tif", package = "SatRbedo") # shortwave-infrared band 2 surface reflectance
+blue_SR <- system.file("extdata/athabasca_2020253_B02_S30.tif", package = "SatRbedo") # blue band surface reflectance (Sentinel-2 band 2)
+green_SR <- system.file("extdata/athabasca_2020253_B03_S30.tif", package = "SatRbedo") # green band surface reflectance (Sentinel-2 band 3)
+red_SR <- system.file("extdata/athabasca_2020253_B04_S30.tif", package = "SatRbedo") # red band surface reflectance (Sentinel-2 band 4)
+NIR_SR <- system.file("extdata/athabasca_2020253_B8A_S30.tif", package = "SatRbedo") # near-infrared band surface reflectance (Sentinel-2 band 8A)
+SWIR1_SR <- system.file("extdata/athabasca_2020253_B11_S30.tif", package = "SatRbedo") # shortwave-infrared band 1 surface reflectance (Sentinel-2 band 11)
+SWIR2_SR <- system.file("extdata/athabasca_2020253_B12_S30.tif", package = "SatRbedo") # shortwave-infrared band 2 surface reflectance (Sentinel-2 band 12)
 
 # Load the DEM and the outline
-# Note: the DEM was re-projected to the extent of the area of interest
+# Note: the DEM was re-projected to the extent of the area of interest and
+# resampled to a 30 m spatial resolution
 dem <- system.file("extdata/athabasca_dem.tif", package = "SatRbedo")
 outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
 ```
@@ -38,10 +40,10 @@ outline <- system.file("extdata/athabasca_outline.shp", package = "SatRbedo")
 ## Step 2: Data pre-processing
 
 ``` r
-# Transform the input data to SpatRaster and crop to the extent of the area of interest
[…]
