### 2025-12-11 a7b5f16
Modified package version and updated references
