+++ NEWS.md
@@ -1,3 +1,3 @@
-# SatRbedo 1.0
+# SatRbedo 1.0.0
 
 * Initial release.
