### 2025-12-12 33098ec
Create CITATION.cff
