### 2026-01-07 #8973 @pabl1t0x
Thank you @ethanwhite, it was a pleasure working with you, too. I really enjoyed the whole experience and learned a lot from you.
