### 2026-01-20 #8973 @pabl1t0x
Hi @kthyng,

Two weeks ago, our paper was accepted for publication, so I am just writing to ask if anything else is needed from my side. Thank you and have a great week.

Pablo
